+++ src/main/java/dinkplugin/BaseNotifier.java
@@ -1,4 +1,4 @@
-package universalDiscord;
+package dinkplugin;
 
 import javax.inject.Inject;
