+++ src/main/java/dinkplugin/LevelNotifier.java
@@ -1,4 +1,4 @@
-package universalDiscord;
+package dinkplugin;
 
 import lombok.extern.slf4j.Slf4j;
