+++ src/main/java/dinkplugin/SlayerNotifier.java
@@ -1,4 +1,4 @@
-package universalDiscord;
+package dinkplugin;
 
 import javax.inject.Inject;
 import java.util.Objects;
