+++ src/main/java/dinkplugin/UniversalDiscordConfig.java
@@ -1,11 +1,11 @@
-package universalDiscord;
+package dinkplugin;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 import net.runelite.client.config.ConfigSection;
 
-@ConfigGroup("universalDiscord")
+@ConfigGroup("dinkplugin")
 public interface UniversalDiscordConfig extends Config {
     @ConfigSection(
             name = "Collection Log",
